fix bug collection view
